Moved collections into util
