+++ src/main/java/fr/estecka/variantscit/ModuleRepository.java
@@ -6,8 +6,8 @@ import java.util.Set;
 import java.util.stream.Stream;
 import org.jetbrains.annotations.Nullable;
 import com.mojang.serialization.DataResult;
-import fr.estecka.variantscit.collections.HashMap2;
-import fr.estecka.variantscit.collections.Map2;
+import fr.estecka.variantscit.util.collections.HashMap2;
+import fr.estecka.variantscit.util.collections.Map2;
 import fr.estecka.variantscit.modules.IBakedModule;
 import fr.estecka.variantscit.modules.cache.CacheBuilder;
 import fr.estecka.variantscit.reload.EModuleHook;
