+++ src/main/java/fr/estecka/variantscit/modules/cache/CacheBuilder.java
@@ -4,9 +4,9 @@ import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
-import fr.estecka.variantscit.collections.HashMap2;
-import fr.estecka.variantscit.collections.HashMap3;
-import fr.estecka.variantscit.collections.Map2;
+import fr.estecka.variantscit.util.collections.HashMap2;
+import fr.estecka.variantscit.util.collections.HashMap3;
+import fr.estecka.variantscit.util.collections.Map2;
 import fr.estecka.variantscit.modules.IBakedModule;
 import fr.estecka.variantscit.modules.IModuleWrapper;
 import fr.estecka.variantscit.modules.ModuleList;
