+++ src/main/java/fr/estecka/variantscit/reload/VariantAggregator.java
@@ -19,9 +19,9 @@ import fr.estecka.variantscit.assetgen.GeneratedResourcePack;
 import fr.estecka.variantscit.assetgen.GeneratorPresets;
 import fr.estecka.variantscit.assetgen.HotswappableResourceManager;
 import fr.estecka.variantscit.assetgen.IAssetGenerator;
-import fr.estecka.variantscit.collections.HashMap2;
-import fr.estecka.variantscit.collections.Map2;
-import fr.estecka.variantscit.collections.NestedMaps;
+import fr.estecka.variantscit.util.collections.HashMap2;
+import fr.estecka.variantscit.util.collections.Map2;
+import fr.estecka.variantscit.util.collections.NestedMaps;
 
 public class VariantAggregator
 {
