+++ src/main/java/fr/estecka/variantscit/util/collections/HashMap3.java
@@ -1,4 +1,4 @@
-package fr.estecka.variantscit.collections;
+package fr.estecka.variantscit.util.collections;
 
 import java.util.HashMap;
 import java.util.Map;
